Implemented API functions and added thunk functions to replace dispatch
This commit has issues! UI is not working properly because of the removal of MainPage
- Added fetchAirconData and sendAirconData to redux/api
- redux/actions no longer handles api requests, axios is only used in api
- Added store to redux/actions so functions there can use store.dispatch() rather than rely on useDispatch from components
- Updated functions from components for the new dispatch method
- Removed client.js (no longer needed)
- TemperatureController component has new updateLabels() method (for testing). Labels update directly from the store rather than from the new values

Things to note:
- useEffect no longer loops because MainPage has been removed and implemented directly to App.js.
The issue was that the way MainPage was implemented, ALL elements and code was being re-created/run every single time there is a change. This has been fixed but now the issue is items from different components are not being updated if a value is changed in one component and the indicator is from another.
- "unexpected end of stream" error from GET Request magically disappeared.

Things to implement:
- Cooldown timer for sendAirconData (don't want to spam server with requests)
- Need to fix component not updating issue.